+++ src/main/java/com/commercetools/sync/services/impl/ProductServiceImpl.java
@@ -1,6 +1,5 @@
 package com.commercetools.sync.services.impl;
 
-import com.commercetools.sync.commons.utils.CtpQueryUtils;
 import com.commercetools.sync.products.ProductSyncOptions;
 import com.commercetools.sync.services.ProductService;
 import io.sphere.sdk.commands.UpdateAction;
@@ -8,29 +7,27 @@ import io.sphere.sdk.products.Product;
 import io.sphere.sdk.products.ProductDraft;
 import io.sphere.sdk.products.commands.ProductCreateCommand;
 import io.sphere.sdk.products.commands.ProductUpdateCommand;
+import io.sphere.sdk.products.expansion.ProductExpansionModel;
 import io.sphere.sdk.products.queries.ProductQuery;
+import io.sphere.sdk.products.queries.ProductQueryModel;
 import io.sphere.sdk.queries.QueryPredicate;
 import org.apache.commons.lang3.StringUtils;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 import static java.util.Collections.singleton;
-import static org.apache.commons.lang3.StringUtils.isBlank;
 
 
-public final class ProductServiceImpl
-    extends BaseService<ProductDraft, Product, ProductSyncOptions> implements ProductService {
+public final class ProductServiceImpl extends BaseService<ProductDraft, Product, ProductSyncOptions, ProductQuery,
+    ProductQueryModel, ProductExpansionModel<Product>> implements ProductService {
 
     public ProductServiceImpl(@Nonnull final ProductSyncOptions syncOptions) {
         super(syncOptions);
@@ -39,41 +36,21 @@ public final class ProductServiceImpl
     @Nonnull
     @Override
     public CompletionStage<Optional<String>> getIdFromCacheOrFetch(@Nullable final String key) {
-        if (isBlank(key)) {
-            return CompletableFuture.completedFuture(Optional.empty());
-        }
-        if (keyToIdCache.containsKey(key)) {
-            return CompletableFuture.completedFuture(Optional.of(keyToIdCache.get(key)));
-        }
-        return fetchAndCache(key);
-    }
 
-    @Nonnull
-    private CompletionStage<Optional<String>> fetchAndCache(@Nonnull final String key) {
-        return cacheKeysToIds(singleton(key)).thenApply(result -> Optional.ofNullable(keyToIdCache.get(key)));
+        return fetchCachedResourceId(key,
+            () -> ProductQuery.of()
+                              .withPredicates(buildProductKeysQueryPredicate(singleton(key))));
     }
 
     @Nonnull
     @Override
     public CompletionStage<Map<String, String>> cacheKeysToIds(@Nonnull final Set<String> productKeys) {
-        final Set<String> keysNotCached = productKeys.stream()
-                                                     .filter(StringUtils::isNotBlank)
-                                                     .filter(key -> !keyToIdCache.containsKey(key))
-                                                     .collect(Collectors.toSet());
 
-        if (keysNotCached.isEmpty()) {
-            return CompletableFuture.completedFuture(keyToIdCache);
-        }
-
-        final ProductQuery productQuery = ProductQuery.of()
-                                                      .withPredicates(buildProductKeysQueryPredicate(keysNotCached));
-
-        return CtpQueryUtils.queryAll(syncOptions.getCtpClient(), productQuery, this::cacheProductIds)
-                            .thenApply(result -> keyToIdCache);
-    }
-
-    private void cacheProductIds(@Nonnull final List<Product> products) {
-        products.forEach(product -> keyToIdCache.put(product.getKey(), product.getId()));
+        return cacheKeysToIds(
+            productKeys,
+            keysNotCached -> ProductQuery
+                .of()
+                .withPredicates(buildProductKeysQueryPredicate(keysNotCached)));
     }
 
     QueryPredicate<Product> buildProductKeysQueryPredicate(@Nonnull final Set<String> productKeys) {
@@ -90,42 +67,18 @@ public final class ProductServiceImpl
     @Nonnull
     @Override
     public CompletionStage<Set<Product>> fetchMatchingProductsByKeys(@Nonnull final Set<String> productKeys) {
-        if (productKeys.isEmpty()) {
-            return CompletableFuture.completedFuture(Collections.emptySet());
-        }
 
-        final ProductQuery productQuery = ProductQuery.of().withPredicates(buildProductKeysQueryPredicate(productKeys));
-
-        return CtpQueryUtils
-                .queryAll(syncOptions.getCtpClient(), productQuery, Function.identity())
-                .thenApply(fetchedProducts -> fetchedProducts
-                        .stream()
-                        .flatMap(List::stream)
-                        .peek(product -> keyToIdCache.put(product.getKey(), product.getId()))
-                        .collect(Collectors.toSet()));
+        return fetchMatchingResources(productKeys,
+            () -> ProductQuery.of().withPredicates(buildProductKeysQueryPredicate(productKeys)));
     }
 
     @Nonnull
     @Override
     public CompletionStage<Optional<Product>> fetchProduct(@Nullable final String key) {
 
-        if (isBlank(key)) {
-            return CompletableFuture.completedFuture(Optional.empty());
-        }
-
-        final ProductQuery productQuery = ProductQuery
-                .of().withPredicates(buildProductKeysQueryPredicate(singleton(key)));
-
-        return syncOptions
-                .getCtpClient()
-                .execute(productQuery)
-                .thenApply(productPagedQueryResult ->
-                        productPagedQueryResult
-                                .head()
-                                .map(product -> {
-                                    keyToIdCache.put(product.getKey(), product.getId());
-                                    return product;
-                                }));
+        return fetchResource(key,
+            () -> ProductQuery
+                .of().withPredicates(buildProductKeysQueryPredicate(singleton(key))));
     }
 
     @Nonnull