+++ docs/RELEASE_NOTES.md
@@ -33,8 +33,19 @@
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/1.6.1...1.7.0) |
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/1.7.0/) | 
 [Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/1.7.0)
--->
 
+
+- ✨ **Enhancements** (2)
+    - **Product Sync** - Only cache the needed keys of `Category`, `ProductType` and `Type` references instead of 
+    caching all keys of such resources. [#418](https://github.com/commercetools/commercetools-sync-java/issues/418) 
+    
+    **Note**: This might have performance implications on the product sync, since now every non cached key-id entry will be individually fetched. 
+    However, issue [#235](https://github.com/commercetools/commercetools-sync-java/issues/235) should address this.
+    
+    - **Commons** - Refactor duplicate implementations in concrete services and generalise it in the `BaseService`. 
+    Thanks, @jarzynp, for the contributions! [#418](https://github.com/commercetools/commercetools-sync-java/issues/418)
+
+-->
 
 ### 1.6.1 -  Oct 17, 2019
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/1.6.0...1.6.1) |